refactor: consolidate duplicate escapeHtml/formatFileSize/getTimeSince into webview/shared
Move three utility functions that were copy-pasted across webview files into
the shared formatUtils module:

- escapeHtml: was duplicated in dashboard, maturity, logviewer, diagnostics,
  fluency-level-viewer, and usage. Canonical version includes all 5 HTML
  entity replacements (including single-quote) for consistent XSS safety.
- formatFileSize: was duplicated in logviewer and diagnostics. Uses the
  more robust diagnostics version with NaN/negative-value guards.
- getTimeSince: was only in diagnostics but belongs in shared for reuse.

All six consumer files now import these from ../shared/formatUtils.
