+++ vscode-extension/src/webview/logviewer/main.ts
@@ -1,6 +1,6 @@
 ﻿// Log Viewer webview - displays session file details and chat turns
 import { ContextReferenceUsage, getTotalContextRefs, getImplicitContextRefs, getExplicitContextRefs, getContextRefsSummary } from '../shared/contextRefUtils';
-import { formatCompact, setCompactNumbers } from '../shared/formatUtils';
+import { escapeHtml, formatCompact, formatFileSize, setCompactNumbers } from '../shared/formatUtils';
 import { getModelDisplayName } from '../shared/modelUtils';
 // CSS imported as text via esbuild
 import themeStyles from '../shared/theme.css';
@@ -170,15 +170,6 @@ return id;
 return TOOL_NAME_MAP[id] ?? TOOL_NAME_MAP[id.toLowerCase()] ?? id;
 }
 
-function escapeHtml(text: string): string {
-return text
-.replace(/&/g, '&amp;')
-.replace(/</g, '&lt;')
-.replace(/>/g, '&gt;')
-.replace(/"/g, '&quot;')
-.replace(/'/g, '&#039;');
-}
-
 function getEffortDisplayName(level: string): string {
 return EFFORT_DISPLAY_NAMES[level] ?? level;
 }
@@ -190,12 +181,6 @@ return new Date(isoString).toLocaleString();
 } catch {
 return isoString;
 }
-}
-
-function formatFileSize(bytes: number): string {
-if (bytes < 1024) { return `${bytes} B`; }
-if (bytes < 1024 * 1024) { return `${(bytes / 1024).toFixed(1)} KB`; }
-return `${(bytes / (1024 * 1024)).toFixed(2)} MB`;
 }
 
 function getContextRefBadges(refs: ContextReferenceUsage): string {
